requirements.txt
plyer==2.1.0
schedule==1.2.0
python-dateutil==2.8.2
tkcalendar==1.6.1
pytz==2024.1
Pillow>=10.0.0
pystray>=0.19.5
